Store updated system properties in cache and restore them on next run
Prior to this, the configuration cache was ignoring the system property
updates. When running from cache, a default value of the property was
used and the build was (potentially) incorrect. In pathological cases
this caused cyclic cache invalidation (if the updated value was
captured as an input).

This commit adds the infrastructure to capture all updated system
properties and to restore their values when running from the cache.
The only limitation is that only configuration-cache supported values
can be used as property keys and values but this should cover the vast
majority of the use cases.

Issue: #18432