Track system properties clearing
Java has a couple of ways to clear system properties and set up new from
scratch. This commit attempts to reapply clearing when running from the
configuration cache.

The one known caveat is "System.properties = ..." syntax in dynamic
Groovy. It uses uncommon bytecode pattern that is hard to intercept, so
this commit doesn't cover it.

Issue: #18432